day2/autoquant-alpha-day2/requirements.txt
numpy>=1.26.0
pandas>=2.1.0
rich>=13.7.0
alpaca-py>=0.20.0
python-dotenv>=1.0.0
pytest>=7.4.0
